requirements.txt
numpy
pandas
Flask
pymongo
scikit-learn

matplotlib
seaborn

python-dotenv
groq